IOS-7293 Fix review 4
